requirements.txt
# YOLOFoodCal - Core Dependencies
# YOLO26 food detection and calorie estimation

# Core
ultralytics>=8.4.0
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.9.0
Pillow>=10.0.0

# Web UI
streamlit>=1.30.0

# Data
pandas>=2.0.0
PyYAML>=6.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
